Test diagram ready

+++ Projects/DataProviderApp/app/src/main/java/com/example/dataproviderapp/dto/requests/CreateGpsDataRequest.kt
@@ -9,5 +9,5 @@ data class CreateGpsDataRequest(
     val longitudeDeg: Double,
     val accuracyM: Float?,
     val speedKmh: Int?,
-    val bearingDeg: Int?
+    val bearingDeg: Float?
 )
