+++ Projects/server/Data/Entities/GPSData.cs
@@ -9,7 +9,7 @@
         public float LongitudeDeg { get; set; }
         public float? AccuracyM { get; set; }
         public uint? SpeedKmh { get; set; }
-        public ushort? BearingDeg { get; set; }
+        public float? BearingDeg { get; set; }
         public Trip Trip { get; set; } = null!;
     }
 }
